+++ lib/util/themes.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class Themes {
+  Themes();
+  final backgroundColor = Colors.grey[850];
+  final primaryColor = Colors.black;
+  final buttonColor = Colors.black;
+  final textColor = Colors.limeAccent[700];
+  final textStyle = new TextStyle(
+    color: Colors.limeAccent[700],
+  );
+}