+++ lib/views/home_page.dart
@@ -19,10 +19,14 @@ class HomePageWidgetState extends State<HomePage> {
 
   HomePageWidgetState(this.onSignedOut);
   final VoidCallback onSignedOut;
-  final _bottomNavigationColor = Colors.grey;
   int _currentIndex = 0;
   List<Widget> pages = List();
   final AuthService auth = new AuthService();
+
+  final _textColor = Colors.limeAccent[700];
+  final _textStyle = new TextStyle(
+    color: Colors.limeAccent[700],
+  );
 
   @override
   void initState() {
@@ -37,74 +41,96 @@ class HomePageWidgetState extends State<HomePage> {
   @override
   Widget build(BuildContext context) {
     return new Scaffold(
-      /*
-      appBar: new AppBar(
-        title: new Text('Welcome'),
-        actions: <Widget>[
-          new FlatButton(
-            child: new Text('Logout', style: new TextStyle(fontSize: 17.0, color: Colors.white)),
-            onPressed: _signedOut,
-        ],
-          )
-      ),
-      */
       body: pages[_currentIndex],
-      bottomNavigationBar: BottomNavigationBar(
-        items: [
-          BottomNavigationBarItem(
-            icon: Icon(
-              Icons.account_circle,
-              color: _bottomNavigationColor,
-            ),
-            title: Text(
-              'Profile',
-              style: TextStyle(color: _bottomNavigationColor),
-            )),
-          BottomNavigationBarItem(
-            icon: Icon(
-              Icons.face,
-              color: _bottomNavigationColor,
-            ),
-            title: Text(
-              'Players',
-              style: TextStyle(color: _bottomNavigationColor),
-            )),
-          BottomNavigationBarItem(
-            icon: Icon(
-              Icons.favorite,
-              color: _bottomNavigationColor,
-            ),
-            title: Text(
-              'Charities',
-              style: TextStyle(color: _bottomNavigationColor),
-            )),
-          BottomNavigationBarItem(
-            icon: Icon(
-              Icons.calendar_today,
-              color: _bottomNavigationColor,
-            ),
-            title: Text(
-              'My Matches',
-              style: TextStyle(color: _bottomNavigationColor),
-            )),
-          BottomNavigationBarItem(
-            icon: Icon(
-              Icons.text_format,
-              color: _bottomNavigationColor
-            ),
-            title: Text(
-              'About',
-              style: TextStyle(color: _bottomNavigationColor),
-            )),
-        ],
-        currentIndex: _currentIndex,
-        onTap: (int index) {
-          setState(() {
-            _currentIndex = index;
-          });
-        },
-        type: BottomNavigationBarType.shifting,
+      bottomNavigationBar: new Theme(
+        data: new ThemeData(
+          canvasColor: Colors.black,
+        ),
+        child: new BottomNavigationBar(
+          items: [
+            BottomNavigationBarItem(
+                icon: Icon(
+                  Icons.account_circle,
+                  color: _textColor,
+                ),
+                title: Text(
+                    'Profile',
+                    style: _textStyle,
+                )),
+            BottomNavigationBarItem(
+                icon: Icon(
+                  Icons.face,
+                  color: _textColor,
+                ),
+                title: Text(
+                  'Players',
+                  style: _textStyle,
+                )),
+            BottomNavigationBarItem(
+                icon: Icon(
+                  Icons.favorite,
+                  color: _textColor,
+                ),
+                title: Text(
+                    'Charities',
+                    style: _textStyle,
+                )),
+            BottomNavigationBarItem(
+                icon: Icon(
+                  Icons.calendar_today,
+                  color: _textColor,
+                ),
+                title: Text(
+                  'My Matches',
+                  style: _textStyle,
+                )),
+            BottomNavigationBarItem(
+                icon: Icon(
+                  Icons.text_format,
+                  color: _textColor,
+                ),
+                title: Text(
+                  'About',
+                  style: _textStyle,
+                )),
+          ],
+          currentIndex: _currentIndex,
+          onTap: (int index) {
+            setState(() {
+              _currentIndex = index;
+            });
+          },
+          type: BottomNavigationBarType.shifting,
+        ),
       ),
     );
   }
 }
+
+/*
+bottomNavigationBar: new Theme(
+  data: Theme.of(context).copyWith(
+    // sets the background color of the `BottomNavigationBar`
+    canvasColor: Colors.green,
+    // sets the active color of the `BottomNavigationBar` if `Brightness` is light
+    primaryColor: Colors.red,
+    textTheme: Theme
+      .of(context)
+      .textTheme
+      .copyWith(caption: new TextStyle(color: Colors.yellow))), // sets the inactive color of the `BottomNavigationBar`
+  child: new BottomNavigationBar(
+    type: BottomNavigationBarType.fixed,
+    currentIndex: 0,
+    items: [
+      new BottomNavigationBarItem(
+        icon: new Icon(Icons.add),
+        title: new Text("Add"),
+      ),
+      new BottomNavigationBarItem(
+        icon: new Icon(Icons.delete),
+        title: new Text("Delete"),
+       )
+      ],
+    ),
+),
+*/